fixed task scheduling and /schedule add (presumably)

+++ src/data/tasks/tasks.py
@@ -36,8 +36,15 @@ class TaskBase:
         else:
             raise e
 
+    @classmethod
+    def runtime_only(cl) -> bool: return False
+
     @abstractclassmethod
     async def execute(cl, obj: object) -> None: pass
+
+    @classmethod
+    def by_type(cl, task_type: TaskExecutionType) -> Type['TaskBase']:
+        return next(taskbase for taskbase in TaskBase._registered_tasks if taskbase.type() == task_type)
 
 
 class Task:
@@ -53,11 +60,11 @@ class Task:
             record = db.query(f'select execution_time, data, task_type from tasks where id={id}')
             if record:
                 self.id = id
-                self.time = record[0]
-                self.data = loads(record[1])
-                for type in TaskBase._registered_tasks:
-                    if type.type().value == record[2]:
-                        self.base = type
+                self.time = record[0][0]
+                self.data = loads(record[0][1])
+                for task_base in TaskBase._registered_tasks:
+                    if task_base.type().value == record[0][2]:
+                        self.base = task_base
                         break
         finally:
             db.disconnect()
@@ -65,6 +72,10 @@ class Task:
     @property
     def type(self) -> TaskExecutionType:
         return self.base.type()
+
+    @property
+    def runtime_only(self) -> bool:
+        return self.base.runtime_only()
 
     async def execute(self) -> None:
         try:
@@ -74,6 +85,7 @@ class Task:
 
 class Tasks:
     _list: List[Task] = []
+    _runtime_tasks: List[Task] = []
 
     def load(self):
         db = bot.instance.data.db
@@ -88,14 +100,28 @@ class Tasks:
         finally:
             db.disconnect()
 
-    def add_task(self, time: datetime, type: TaskExecutionType, data: object = None) -> None:
-        db = bot.instance.data.db
-        db.connect()
-        try:
-            db.query(f'insert into tasks (execution_time, task_type, data) values ({pg_timestamp(time)}, {type.value}, \'{dumps(data)}\') returning id')
-            self.load()
-        finally:
-            db.disconnect()
+    def sort_runtime_tasks_by_time(self) -> None:
+        self._runtime_tasks = sorted(self._runtime_tasks, key=lambda task: task.time)
+
+    def add_task(self, time: datetime, task_type: TaskExecutionType, data: object = None) -> None:
+        if TaskBase.by_type(task_type).runtime_only():
+            task = Task()
+            task.time = time
+            task.data = data
+            for task_base in TaskBase._registered_tasks:
+                if task_base.type() == task_type:
+                    task.base = task_base
+                    break
+            self._runtime_tasks.append(task)
+            self.sort_runtime_tasks_by_time()
+        else:
+            db = bot.instance.data.db
+            db.connect()
+            try:
+                db.query(f'insert into tasks (execution_time, task_type, data) values ({pg_timestamp(time)}, {task_type.value}, \'{dumps(data)}\') returning id')
+                self.load()
+            finally:
+                db.disconnect()
 
     def contains(self, type: TaskExecutionType) -> bool:
         for task in self._list:
@@ -105,36 +131,60 @@ class Tasks:
 
     def get_next(self) -> Task:
         """Gets the next possible task to be executed."""
+        task1, task2 = None, None
         if self._list and self._list[0].time <= datetime.utcnow():
-            return self._list[0]
+            task1 = self._list[0]
+        if self._runtime_tasks and self._runtime_tasks[0].time <= datetime.utcnow():
+            task2 = self._runtime_tasks[0]
+
+        if task1 and task2:
+            if task1.time < task2.time:
+                return task1
+            else:
+                return task2
+        elif task1:
+            return task1
+        elif task2:
+            return task2
+
         return None
 
-    def remove_task(self, id: int):
-        db = bot.instance.data.db
-        db.connect()
-        try:
-            db.query(f'delete from tasks where id={id}')
-            self.load()
-        finally:
-            db.disconnect()
+    def remove_task(self, task: Task):
+        if task.runtime_only:
+            if task in self._runtime_tasks:
+                self._runtime_tasks.remove(task)
+        else:
+            db = bot.instance.data.db
+            db.connect()
+            try:
+                db.query(f'delete from tasks where id={task.id}')
+                self.load()
+            finally:
+                db.disconnect()
 
     def remove_all(self, type: TaskExecutionType):
-        db = bot.instance.data.db
-        db.connect()
-        try:
-            db.query(f'delete from tasks where task_type={type.value}')
-            self.load()
-        finally:
-            db.disconnect()
+        if TaskBase.by_type(type).runtime_only():
+            self._runtime_tasks = list(filter(lambda task: task.type == type, self._runtime_tasks))
+        else:
+            db = bot.instance.data.db
+            db.connect()
+            try:
+                db.query(f'delete from tasks where task_type={type.value}')
+                self.load()
+            finally:
+                db.disconnect()
 
     def remove_task_by_data(self, type: TaskExecutionType, data: object):
         if data is None:
             return
 
-        db = bot.instance.data.db
-        db.connect()
-        try:
-            db.query(f'delete from tasks where task_type={type.value} and data=\'{dumps(data)}\'')
-            self.load()
-        finally:
-            db.disconnect()
+        if TaskBase.by_type(type).runtime_only():
+            self._runtime_tasks = list(filter(lambda task: task.type == type and task.data == data, self._runtime_tasks))
+        else:
+            db = bot.instance.data.db
+            db.connect()
+            try:
+                db.query(f'delete from tasks where task_type={type.value} and data=\'{dumps(data)}\'')
+                self.load()
+            finally:
+                db.disconnect()
